backend/requirements.txt
numpy
Pillow
asgiref
django
djangorestframework
django-ninja
djangorestframework-simplejwt

libsass 
django-compressor 
django-sass-processor
pytz


pymongo
django-seed
gunicorn
psycopg2-binary
sqlparse
mysqlclient
mongoengine
argon2-cffi